add vendor caracterists and item for them

+++ app/Http/Controllers/Backend/VendorProductImageGalleryController.php
@@ -8,6 +8,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\ProductImageGallery;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class VendorProductImageGalleryController extends Controller
 {
@@ -19,6 +20,10 @@ class VendorProductImageGalleryController extends Controller
     public function index(VendorProductImageGalleryDataTable $dataTable, Request $request)
     {
         $product = Product::findOrFail($request->product);
+
+        if($product->vendor_id !== Auth::user()->vendor->id){
+            abort(404);
+        }
 
         return $dataTable->render('vendor.product.image-gallery.index', compact('product'));
     }
@@ -87,6 +92,10 @@ class VendorProductImageGalleryController extends Controller
     public function destroy(string $id)
     {
         $galleryImage = ProductImageGallery::findOrFail($id);
+
+        if($galleryImage->product->vendor->id !== Auth::user()->vendor->id){
+            abort(404);
+        }
 
         $this->deleteImage($galleryImage->image);
